Return a GPUSet from `LearnerTrainParam`.
