+++ src/common/common.cu
@@ -1,7 +1,8 @@
 /*!
- * Copyright 2018 XGBoost contributors
+ * Copyright 2018-2019 XGBoost contributors
  */
 #include "common.h"
+#include "xgboost/gpuset.h"
 
 namespace xgboost {
