+++ src/common/host_device_vector.h
@@ -60,6 +60,7 @@
 #include <utility>
 #include <vector>
 
+#include "xgboost/gpuset.h"
 #include "common.h"
 #include "span.h"
