+++ tests/cpp/common/test_common.cc
@@ -1,5 +1,9 @@
-#include "../../../src/common/common.h"
+/*!
+ * Copyright 2018-2019 XGBoost contributors
+ */
 #include <gtest/gtest.h>
+#include "xgboost/gpuset.h"
+#include "../../../src/common/common.h"
 
 namespace xgboost {
 TEST(GPUSet, Basic) {
