+++ tests/cpp/common/test_common.cu
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 #include "xgboost/logging.h"
+#include "xgboost/gpuset.h"
 #include "../../../src/common/common.h"
 #include "../helpers.h"
